# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ecommerce de productos orgánicos",
        "description": "Documentación para la API de ecommerce de productos orgánicos",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend-raices.vercel.app/api/api",
            "description": "Servidor de producción"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/cart/products": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Agregar un producto al carrito",
                "operationId": "7b18094197fb0278033a7d005defd910",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID del producto",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Cantidad del producto",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto añadido al carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de éxito",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stock insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Eliminar un producto del carrito",
                "operationId": "3abba5fe9fb37d322eef8640dec89cba",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID del producto a eliminar",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto eliminado del carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de éxito",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado en el carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Cart"
                ],
                "summary": "Actualizar cantidad de un producto en el carrito",
                "operationId": "b78cf420f8f8627ebcbb0215db342b56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "product_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "product_id": {
                                        "description": "ID del producto",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Nueva cantidad del producto",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cantidad actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de éxito",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado en el carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Stock insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Ver productos en el carrito",
                "operationId": "d0af7f3dc67d69cf8fbf09382808f844",
                "responses": {
                    "200": {
                        "description": "Lista de productos en el carrito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cartItems": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CartItem"
                                            }
                                        },
                                        "totalPrice": {
                                            "description": "Precio total del carrito",
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/cart/checkout": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Confirmar compra y crear un pedido",
                "operationId": "a2ca04e6b844f8dfdba0ca974f51a791",
                "responses": {
                    "200": {
                        "description": "Pedido realizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de éxito",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Carrito vacío o stock insuficiente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Mensaje de error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CartItem": {
                "properties": {
                    "product_id": {
                        "description": "ID del producto",
                        "type": "integer"
                    },
                    "quantity": {
                        "description": "Cantidad del producto en el carrito",
                        "type": "integer"
                    },
                    "product": {
                        "description": "Detalles del producto",
                        "properties": {
                            "id": {
                                "description": "ID del producto",
                                "type": "integer"
                            },
                            "name": {
                                "description": "Nombre del producto",
                                "type": "string"
                            },
                            "price": {
                                "description": "Precio del producto",
                                "type": "number",
                                "format": "float"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}